+++ src/main/scala/co/topl/attestation/Proof.scala
@@ -1,6 +1,7 @@
 package co.topl.attestation
 
 import co.topl.attestation.serialization.ProofSerializer
+import co.topl.crypto.{PrivateKeyCurve25519, Secret}
 import co.topl.utils.serialization.{BifrostSerializer, BytesSerializable}
 import com.google.common.primitives.Ints
 import io.circe.syntax.EncoderOps
