+++ src/main/scala/co/topl/attestation/Proposition.scala
@@ -4,6 +4,7 @@ import co.topl.attestation.AddressEncoder.NetworkPrefix
 import co.topl.attestation.Evidence.{EvidenceContent, EvidenceTypePrefix}
 import co.topl.attestation.PublicKeyPropositionCurve25519.{typePrefix, typeString}
 import co.topl.attestation.serialization.PropositionSerializer
+import co.topl.crypto.{PrivateKeyCurve25519, Secret}
 import co.topl.utils.{Identifiable, Identifier}
 import co.topl.utils.serialization.{BifrostSerializer, BytesSerializable}
 import com.google.common.primitives.Ints
