+++ src/main/scala/co/topl/consensus/genesis/PrivateTestnet.scala
@@ -7,10 +7,8 @@ import co.topl.consensus.Forger.ChainParams
 import co.topl.modifier.ModifierId
 import co.topl.modifier.block.Block
 import co.topl.modifier.transaction.{ArbitTransfer, PolyTransfer}
-import co.topl.nodeView.history.History
 import co.topl.nodeView.state.box.{ArbitBox, SimpleValue}
-import co.topl.settings.{AppSettings, RuntimeOpts, Version}
-import co.topl.utils.encode.encodeBase16
+import co.topl.settings.{AppSettings, Version}
 
 import scala.util.Try
