+++ src/main/scala/co/topl/crypto/PrivateKeyCurve25519.scala
@@ -1,6 +1,6 @@
-package co.topl.attestation
+package co.topl.crypto
 
-import co.topl.crypto.KeyfileCurve25519
+import co.topl.attestation.{PublicKeyPropositionCurve25519, SignatureCurve25519}
 import co.topl.utils.serialization.{BifrostSerializer, Reader, Writer}
 import scorex.crypto.signatures.{Curve25519, PrivateKey, PublicKey}
