+++ src/main/scala/co/topl/modifier/block/serialization/BlockBodySerializer.scala
@@ -4,7 +4,7 @@ import co.topl.modifier.ModifierId
 import co.topl.modifier.block.BlockBody
 import co.topl.modifier.transaction.Transaction
 import co.topl.modifier.transaction.serialization.TransactionSerializer
-import co.topl.utils.Extensions.LongOps
+import co.topl.utils.serialization.Extensions.LongOps
 import co.topl.utils.serialization.{BifrostSerializer, Reader, Writer}
 
 object BlockBodySerializer extends BifrostSerializer[BlockBody] {
