+++ src/main/scala/co/topl/modifier/transaction/serialization/ArbitTransferSerializer.scala
@@ -4,7 +4,7 @@ import co.topl.attestation._
 import co.topl.attestation.serialization.{ProofSerializer, PropositionSerializer}
 import co.topl.modifier.transaction.ArbitTransfer
 import co.topl.nodeView.state.box.TokenValueHolder
-import co.topl.utils.Extensions._
+import co.topl.utils.serialization.Extensions._
 import co.topl.utils.serialization.{BifrostSerializer, Reader, Writer}
 
 import scala.language.existentials
