+++ src/main/scala/co/topl/network/peer/PeerSpecSerializer.scala
@@ -3,7 +3,7 @@ package co.topl.network.peer
 import java.net.{InetAddress, InetSocketAddress}
 
 import co.topl.settings.{Version, VersionSerializer}
-import co.topl.utils.Extensions._
+import co.topl.utils.serialization.Extensions._
 import co.topl.utils.serialization.{BifrostSerializer, Reader, Writer}
 
 class PeerSpecSerializer(featureSerializers: PeerFeature.Serializers) extends BifrostSerializer[PeerSpec] {
