+++ src/main/scala/co/topl/nodeView/state/box/serialization/ExecutionBoxSerializer.scala
@@ -1,7 +1,7 @@
 package co.topl.nodeView.state.box.serialization
 
 import co.topl.nodeView.state.box.{ExecutionBox, ProgramId}
-import co.topl.utils.Extensions._
+import co.topl.utils.serialization.Extensions._
 import co.topl.utils.serialization.{BifrostSerializer, Reader, Writer}
 
 object ExecutionBoxSerializer extends BifrostSerializer[ExecutionBox] {
