+++ src/main/scala/co/topl/utils/serialization/BifrostSerializer.scala
@@ -1,9 +1,8 @@
 package co.topl.utils.serialization
 
 import akka.util.ByteString
-import co.topl.utils.ByteArrayBuilder
-
 import java.nio.ByteBuffer
+
 import scala.util.Try
 
 trait BifrostSerializer[Content] extends Serializer[Content, Content, Reader, Writer] {
