+++ src/main/scala/co/topl/utils/serialization/VLQWriter.scala
@@ -1,6 +1,6 @@
 package co.topl.utils.serialization
 
-import co.topl.utils.encode.ZigZagEncoder._
+import ZigZagEncoder._
 
 import java.util
