+++ src/test/scala/co/topl/crypto/Bip39Spec.scala
@@ -1,6 +1,5 @@
 package co.topl.crypto
 
-import co.topl.attestation.PrivateKeyCurve25519
 import co.topl.utils.Logging
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
