+++ src/test/scala/co/topl/utils/GenesisGenerators.scala
@@ -1,9 +1,8 @@
 package co.topl.utils
 
-import co.topl.attestation.PrivateKeyCurve25519
 import co.topl.consensus.KeyRing
 import co.topl.consensus.genesis.PrivateTestnet
-import co.topl.crypto.KeyfileCurve25519
+import co.topl.crypto.{KeyfileCurve25519, PrivateKeyCurve25519}
 import co.topl.modifier.ModifierId
 import co.topl.modifier.block.Block
 import co.topl.nodeView.history.History
