+++ src/test/scala/co/topl/utils/ValidGenerators.scala
@@ -1,10 +1,10 @@
 package co.topl.utils
 
-import co.topl.attestation.{PrivateKeyCurve25519, PublicKeyPropositionCurve25519}
+import co.topl.attestation.PublicKeyPropositionCurve25519
 import co.topl.attestation.PublicKeyPropositionCurve25519.evProducer
 import co.topl.consensus.KeyRing
 import co.topl.consensus.genesis.PrivateTestnet
-import co.topl.crypto.KeyfileCurve25519
+import co.topl.crypto.{KeyfileCurve25519, PrivateKeyCurve25519}
 import co.topl.modifier.ModifierId
 import co.topl.modifier.block.Block
 import co.topl.modifier.transaction.Transaction.TX
